가입 시 권한 관련 오류 수정 및 Like 관련 테스트 추가 작성

+++ sns/src/main/java/com/project/sns/repository/AuthorityRepository.java
@@ -14,6 +14,6 @@ public class AuthorityRepository {
     private final EntityManager em;
 
     public Authority getUserAuthority() {
-        return em.find(Authority.class, 2);
+        return em.find(Authority.class, (long) 2);
     }
 }
